+++ packages/pluggableWidgets/tree-node-web/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- We added icons for dark mode.
+- We added dark mode to Structure mode preview.
+
 ## [1.0.2] - 2021-12-3
 
 ### Fixed